+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- instance: enable tpm cmd
+- instance: secureboot and tpm support on create
 - sks: add rotate operators CA cmd
 - sks: add rotate CSI credentials cmd
 - sks: feature-gates support for cluster update and creation #677
@@ -12,6 +14,7 @@
 
 ### Improvements
 
+- instance: show migrated to egoscale v3
 - SKS: more related commands are migrated to egoscale v3
 - dbaas: move all commands to egoscale v3
 - EIP: move to egoscale v3 #684